delete local copy only if the operation being reverted is actually a rename

+++ plugins/svn4idea/src/org/jetbrains/idea/svn/checkin/SvnCheckinEnvironment.java
@@ -284,7 +284,9 @@ public class SvnCheckinEnvironment implements CheckinEnvironment {
         if (!afterPath.equals(beforePath)) {
           checkRevertFile(afterPath, exceptions);
           // rolling back a rename should delete the after file
-          FileUtil.delete(afterPath);
+          if (beforePath != null) {
+            FileUtil.delete(afterPath);
+          }
         }
       }
     }
